Use search controller in SCPageSearch.
Issue #3536.
