--- client/controllers/sc-search-controller.js
@@ -1,3 +1,0 @@
-export class SCSearchController {
-
-}
