+++ client/elements/sc-search-controller.js
@@ -1,0 +1,13 @@
+export class SCSearchController {
+  async fetchResult(requestUrl, selectedLanguages) {
+    return await (
+      await fetch(requestUrl, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(selectedLanguages),
+      })
+    ).json();
+  }
+}
